Create new 'filehashstore_exceptions' module and refactor filehashstore & update pytests
