+++ src/hashstore/filehashstore.py
@@ -14,6 +14,20 @@ from tempfile import NamedTemporaryFile
 import fcntl
 import yaml
 from hashstore import HashStore, ObjectMetadata
+from hashstore.filehashstore_exceptions import (
+    CidRefsContentError,
+    CidRefsDoesNotExist,
+    CidRefsFileNotFound,
+    NonMatchingChecksum,
+    NonMatchingObjSize,
+    PidAlreadyExistsError,
+    PidNotFoundInCidRefsFile,
+    PidRefsContentError,
+    PidRefsDoesNotExist,
+    PidRefsFileNotFound,
+    RefsFileExistsButCidObjMissing,
+    UnsupportedAlgorithm,
+)
 
 
 class FileHashStore(HashStore):
@@ -2570,110 +2584,3 @@ class Stream(object):
             self._obj.close()
         else:
             self._obj.seek(self._pos)
-
-
-class PidRefsFileNotFound(Exception):
-    """Custom exception thrown when verifying reference files and a pid refs
-    file is not found."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class CidRefsFileNotFound(Exception):
-    """Custom exception thrown when verifying reference files and a cid refs
-    file is not found."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class PidRefsContentError(Exception):
-    """Custom exception thrown when verifying reference files and a pid refs
-    file does not contain the cid that it is expected."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class CidRefsContentError(Exception):
-    """Custom exception thrown when verifying reference files and a cid refs
-    file does not have a pid that is expected to be found."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class PidAlreadyExistsError(Exception):
-    """Custom exception thrown when a client calls 'tag_object' and the pid
-    that is being tagged is already accounted for (has a pid refs file and
-    is found in the cid refs file)."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class PidRefsDoesNotExist(Exception):
-    """Custom exception thrown when a pid refs file does not exist."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class CidRefsDoesNotExist(Exception):
-    """Custom exception thrown when a cid refs file does not exist."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class RefsFileExistsButCidObjMissing(Exception):
-    """Custom exception thrown when pid and cid refs file exists, but the
-    cid object does not."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class PidNotFoundInCidRefsFile(Exception):
-    """Custom exception thrown when pid reference file exists with a cid, but
-    the respective cid reference file does not contain the pid."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class NonMatchingObjSize(Exception):
-    """Custom exception thrown when verifying an object and the expected file size
-    does not match what has been calculated."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class NonMatchingChecksum(Exception):
-    """Custom exception thrown when verifying an object and the expected checksum
-    does not match what has been calculated."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
-
-
-class UnsupportedAlgorithm(Exception):
-    """Custom exception thrown when a given algorithm is not supported in HashStore for
-    calculating hashes/checksums, as the default store algo and/or other operations."""
-
-    def __init__(self, message, errors=None):
-        super().__init__(message)
-        self.errors = errors
